requirements.txt
albumentations==1.0.3
click==8.0.3
clip-by-openai==1.1
einops==0.3.0
glfw==2.5.0
gradio==2.3.9
imageio==2.9.0
imgui==1.4.1
kornia==0.5.10
lmdb==0.98
lpips==0.1.4
matplotlib==3.4.3
numpy==1.21.2
hydra-core==1.1
opencv_python_headless==4.5.1.48
Pillow==9.0.1
psutil==5.8.0
PyMCubes==0.1.2
PyOpenGL==3.1.6
pyspng==0.1.0
requests==2.26.0
scipy==1.7.1
submitit==1.1.5
tensorboardX==2.5
tqdm==4.62.2
trimesh==3.9.8
imageio-ffmpeg==0.4.5
torchvision
gdown
ninja
av